+++ projects/packages/videopress/src/class-admin-ui.php
@@ -169,7 +169,7 @@ class Admin_UI {
 	 * @return string
 	 */
 	public static function render_initial_state() {
-		return 'var jetpackVideoPressInitialState=JSON.parse(decodeURIComponent("' . rawurlencode( wp_json_encode( self::initial_state() ) ) . '"));';
+		return 'var jetpackVideoPressInitialState=' . wp_json_encode( self::initial_state(), JSON_UNESCAPED_SLASHES | JSON_HEX_TAG | JSON_HEX_AMP ) . ';';
 	}
 
 	/**