+++ projects/plugins/wpcomsh/wpcomsh.php
@@ -533,7 +533,7 @@ function wpcom_hide_scan_threats_from_api( $response ) {
 	}
 
 	$json_body['threats']  = array();
-	$response_data['data'] = wp_json_encode( $json_body );
+	$response_data['data'] = wp_json_encode( $json_body, JSON_UNESCAPED_SLASHES );
 	$response->set_data( $response_data );
 
 	return $response;
@@ -613,7 +613,7 @@ function wpcomsh_footer_rum_js() {
 	$rum_kv['wptheme_is_block'] = wp_is_block_theme() ? '1' : '0';
 
 	if ( count( $rum_kv ) > 0 ) {
-		$rum_kv = wp_json_encode( $rum_kv, JSON_FORCE_OBJECT );
+		$rum_kv = wp_json_encode( $rum_kv, JSON_FORCE_OBJECT | JSON_UNESCAPED_SLASHES | JSON_HEX_AMP );
 		if ( is_string( $rum_kv ) ) {
 			$rum_kv = 'data-customproperties="' . esc_attr( $rum_kv ) . '"';
 		} else {