+++ projects/packages/search/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.54.6] - 2025-12-01
+### Changed
+- Update package dependencies. [#46143]
+
 ## [0.54.5] - 2025-11-21
 ### Fixed
 - Phan: Address PhanPossiblyUndeclaredVariable violations. [#45911]
@@ -1394,6 +1398,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated package dependencies.
 - Update PHPUnit configs to include just what needs coverage rather than include everything then try to exclude stuff that doesn't.
 
+[0.54.6]: https://github.com/Automattic/jetpack-search/compare/v0.54.5...v0.54.6
 [0.54.5]: https://github.com/Automattic/jetpack-search/compare/v0.54.4...v0.54.5
 [0.54.4]: https://github.com/Automattic/jetpack-search/compare/v0.54.3...v0.54.4
 [0.54.3]: https://github.com/Automattic/jetpack-search/compare/v0.54.2...v0.54.3