+++ projects/packages/videopress/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.32.17] - 2025-12-01
+### Changed
+- Update package dependencies. [#46072] [#46143]
+
 ## [0.32.16] - 2025-11-21
 ### Changed
 - Update package dependencies. [#46022]
@@ -1194,7 +1198,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - VideoPress: add Cancel button to uploading file component when replacing file [#28188]
 - VideoPress: add Replace control to video block [#28162]
 - VideoPress: minor TS enhancement in the useSearchParams() hook [#28250]
-- VideoPress: re-implemnt useResumableUploader(). 
+- VideoPress: re-implemnt useResumableUploader().
   VideoPress: Iterate over resumable file uploader
     * Re implement useResumableUploader() hook with TS
     * Update VideoPress uploader to use this hook
@@ -1788,6 +1792,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Created empty package [#24952]
 
+[0.32.17]: https://github.com/Automattic/jetpack-videopress/compare/v0.32.16...v0.32.17
 [0.32.16]: https://github.com/Automattic/jetpack-videopress/compare/v0.32.15...v0.32.16
 [0.32.15]: https://github.com/Automattic/jetpack-videopress/compare/v0.32.14...v0.32.15
 [0.32.14]: https://github.com/Automattic/jetpack-videopress/compare/v0.32.13...v0.32.14